etnaviv: add _not_ fully working etnaviv DRM
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>

For the moment we only support one drm_vivante_gem_submit_cmd per
submit ioctl. The current kernel interface wants it and I am not
sure if I need to send more then one cmd per submit ioctl. So
I will keep it as it is now and we will see.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>