etnaviv: rework to support recent kernel changes (rmk/pengutronix)
In the kernel we have many gpu cores supporting on
or more pipes. To reflect this change this commit adds
struct etna_gpu, which represents - yes - the gpu core
with its etna_specs. From here we can use etna_pipe to
work with the gpu core provided pipe. The simple
2d rendering test works fine, but assumes that the 2d
capable core is core 0.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>